+++ .github/workflows/create-release-branch.yml
@@ -137,19 +137,9 @@ jobs:
           git push -d origin ${NEW_VERSION}-interim
           # Output the release branch name so we can use it in later jobs.
           echo "branch_name=release-$NEW_VERSION" >> "$GITHUB_OUTPUT"
-  # Build the tarballs so they can be installed locally when we run tests.
-  build-v4-scanner-tarball:
-    name: 'Build v4 scanner tarball'
-    needs: verify-should-run
-    uses: ./.github/workflows/build-tarball.yml
-    with:
-      # Note: Using `dev-4` here is technically incorrect. For full completeness's sake, we should probably be
-      # using the branch corresponding to the upcoming scanner release. However, identifying that branch is
-      # non-trivial, and there are unlikely to be major differences between the two that appear in the few days
-      # between creating the branch and releasing it, so it _should_ be fine.
-      target-branch: 'dev-4'
-  build-v5-code-analyzer-tarball:
-    name: 'Build v5 code-analyzer tarball'
+  # Build the tarball so it can be installed locally when we run tests.
+  build-code-analyzer-tarball:
+    name: 'Build code-analyzer tarball'
     needs: verify-should-run
     uses: ./.github/workflows/build-tarball.yml
     with:
@@ -161,12 +151,11 @@ jobs:
   # Run all the various tests against the newly created branch.
   test-release-branch:
     name: 'Run unit tests'
-    needs: [build-v4-scanner-tarball, build-v5-code-analyzer-tarball, create-release-branch]
+    needs: [build-code-analyzer-tarball, create-release-branch]
     uses: ./.github/workflows/run-tests.yml
     with:
       # We want to validate the extension against whatever version of code analyzer we *plan* to publish,
       # not what's *already* published.
       use-tarballs: true
-      v4-tarball-suffix: 'dev-4'
-      v5-tarball-suffix: 'dev'
+      tarball-suffix: 'dev'
       target-branch: ${{ needs.create-release-branch.outputs.branch-name }}
