+++ .github/workflows/validate-pr.yml
@@ -47,27 +47,21 @@ jobs:
           echo "Valid PR title: '$title'"
   # RUN TESTS
   # Step 1: Build the tarball so it can be installed locally.
-  build_v4_scanner_tarball:
-    name: 'Build v4 scanner tarball'
-    uses: ./.github/workflows/build-tarball.yml
-    with:
-      target-branch: 'dev-4'
-  build_v5_code_analyzer_tarball:
-    name: 'Build v5 code analyzer tarball'
+  build_code_analyzer_tarball:
+    name: 'Build code analyzer tarball'
     uses: ./.github/workflows/build-tarball.yml
     with:
       target-branch: 'dev'
   # Step 2: Actually run the tests.
   run_tests:
     name: 'Run unit tests'
-    needs: [build_v4_scanner_tarball, build_v5_code_analyzer_tarball]
+    needs: [build_code_analyzer_tarball]
     uses: ./.github/workflows/run-tests.yml
     with:
       # We want to validate the extension against whatever version of code analyzer we
       # *plan* to publish, not what's *already* published.
       use-tarballs: true
-      v4-tarball-suffix: 'dev-4'
-      v5-tarball-suffix: 'dev'
+      tarball-suffix: 'dev'
   # BUILD A VSIX ARTIFACT
   # Additionally, build a VSIX that can be downloaded by the user if needed.
   create-vsix-artifact:
