+++ src/lib/code-analyzer-run-action.ts
@@ -59,7 +59,7 @@ export class CodeAnalyzerRunAction {
                     message: messages.scanProgressReport.analyzingTargets,
                     increment: 20
                 });
-                this.logger.log(messages.info.scanningWith(await this.codeAnalyzer.getScannerName()));
+                this.logger.log(messages.info.scanningWith(await this.codeAnalyzer.getVersion()));
                 const violations: Violation[] = await this.codeAnalyzer.scan(workspace);
 
                 progressReporter.reportProgress({
