--- src/lib/scanner-strategies/scanner-strategy.ts
@@ -1,10 +1,0 @@
-import {Violation} from '../diagnostics';
-import {Workspace} from "../workspace";
-
-export interface CliScannerStrategy {
-    scan(workspace: Workspace): Promise<Violation[]>;
-
-    getScannerName(): Promise<string>;
-
-    getRuleDescriptionFor(engineName: string, ruleName: string): Promise<string>;
-}
